Move toast and locale stores to $lib/stores
